+++ src/onegov.ticket/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+0.0.3 (2015-07-15)
+~~~~~~~~~~~~~~~~~~~
+
 - Adds an email property to the handler.
   [href]